[BUGFIX] fix localization for most BE modules
This commit fixes every BE module except for the "SEO" one. Due to a
typo in the Context configuration the values for each page are always
from the default language. The "SEO" module action works as the
Context is already created in the controller action by creating a
TSFE object.